Crate monitor

The Monitor is a simple Discord bot that hangs out in the Narwhals server. The bot offers several chat games as well as tools to automatically perform certain administrative tasks in order to better keep the server organized.

You can add the bot to your own server here, or you can build your own by following these instructions.

Modules

commands
Contains bot commands called on Discord or Matrix.

discord
Contains the main logic relating to the Discord bot.

errors
Contains error types.

Constants

MAX_THREADS
The maximum number of threads allowed to be running simultaneously.