monitor::mon::compartment::runcomp

Struct RunComp

pub struct RunComp {

    pub sctx: ObjID,
    pub instance: ObjID,
    pub name: String,
    pub compartment_id: CompartmentId,
    main: Option<CompThread>,
    pub deps: Vec<ObjID>,
    comp_config_object: CompConfigObject,
    alloc: Talc<ErrOnOom>,
    mapped_objects: HashMap<MapInfo, Arc<MapHandleInner>>,
    flags: Box<AtomicU64>,
    per_thread: HashMap<ObjID, PerThread>,
    init_info: Option<(StackObject, usize, Vec<CtorSet>)>,
    pub(crate) use_count: u64,

}

A runnable or running compartment.

Fields
sctx: ObjID

The security context for this compartment.

instance: ObjID

The instance of the security context.

name: String

The name of this compartment.

compartment_id: CompartmentId

The dynlink ID of this compartment.

main: Option<CompThread>deps: Vec<ObjID>comp_config_object: CompConfigObjectalloc: Talc<ErrOnOom>mapped_objects: HashMap<MapInfo, Arc<MapHandleInner>>flags: Box<AtomicU64>per_thread: HashMap<ObjID, PerThread>init_info: Option<(StackObject, usize, Vec<CtorSet>)>use_count: u64
Implementations

impl RunComp

pub fn new( sctx: ObjID, instance: ObjID, name: String, compartment_id: CompartmentId, deps: Vec<ObjID>, comp_config_object: CompConfigObject, flags: u64, main_stack: StackObject, entry: usize, ctors: &[CtorSet], ) -> Self

Build a new runtime compartment.

pub fn get_per_thread(&mut self, id: ObjID) -> &mut PerThread

Get per-thread data in this compartment.

pub fn clean_per_thread_data(&mut self, id: ObjID)

Remove all per-thread data for a given thread.

pub fn map_object( &mut self, info: MapInfo, handle: Arc<MapHandleInner>, ) -> Result<Arc<MapHandleInner>, TwzError>

Map an object into this compartment.

pub fn unmap_object(&mut self, info: MapInfo) -> Option<Arc<MapHandleInner>>

Unmap and object from this compartment.

pub fn comp_config_ptr(&self) -> *const SharedCompConfig

Get a pointer to the compartment config.

pub fn monitor_new<T: Copy + Sized>(&mut self, data: T) -> Result<*mut T, ()>

Allocate some space in the compartment allocator, and initialize it.

pub fn monitor_new_slice<T: Copy + Sized>( &mut self, data: &[T], ) -> Result<*mut T, ()>

Allocate some space in the compartment allocator for a slice, and initialize it.

pub fn set_flag(&self, val: u64)

Set a flag on this compartment, and wakeup anyone waiting on flag change.

pub fn cas_flag(&self, old: u64, new: u64) -> Result<u64, u64>

Set a flag on this compartment, and wakeup anyone waiting on flag change.

pub fn notify_state_changed(&self)

pub fn has_flag(&self, flag: u64) -> bool

Check if a flag is set.

pub fn until_change(&self, cur: u64) -> ThreadSyncSleep

Setup a ThreadSyncSleep for waiting until the flag is set. Returns None if the flag is already set.

pub fn raw_flags(&self) -> u64

Get the raw flags bits for this RC.

pub(crate) fn start_main_thread( &mut self, state: u64, tmgr: &mut ThreadMgr, dynlink: &mut Context, args: &[&CStr], env: &[&CStr], ) -> Option<bool>

fn build_tls_template(&mut self, dynlink: &mut Context) -> Option<()>

pub fn read_error_code(&self) -> u64

pub(crate) fn inc_use_count(&mut self)

pub(crate) fn dec_use_count(&mut self) -> bool