monitor::mon

Module space

Re-exports

pub use handle::MapHandle;

pub use unmapper::Unmapper;

Modules

handle 🔒

unmapper 🔒

Structs

MapInfo

A mapping of an object and flags.

MappedObject 🔒

Space

An address space we can map objects into.

UnmapOnDrop 🔒

Allows us to call handle_drop and do all the hard work in the caller, since the caller probably had to hold a lock to call these functions.

Functions

__monitor_get_slot 🔒 ⚠

__monitor_get_slot_pair 🔒 ⚠

__monitor_release_pair 🔒 ⚠

__monitor_release_slot 🔒 ⚠

early_object_map

Map an object into the address space, without tracking it. This leaks the mapping, but is useful for bootstrapping. See the object mapping gate comments for more details.

mapflags_into_prot 🔒