monitor::mon

Struct Monitor

pub struct Monitor {
    locks: LockCollection<(&'static RwLock<ThreadMgr>, &'static RwLock<CompartmentMgr>, &'static RwLock<&'static mut Context>, &'static RwLock<HandleMgr<LibraryHandle>>, &'static RwLock<HandleMgr<CompartmentHandle>>)>,
    unmapper: OnceLock<Unmapper>,
    pub space: &'static Mutex<Space>,
    pub thread_mgr: &'static RwLock<ThreadMgr>,
    pub comp_mgr: &'static RwLock<CompartmentMgr>,
    pub dynlink: &'static RwLock<&'static mut Context>,
    pub library_handles: &'static RwLock<HandleMgr<LibraryHandle>>,
    pub _compartment_handles: &'static RwLock<HandleMgr<CompartmentHandle>>,
}

A security monitor instance. All monitor logic is implemented as methods for this type. We split the state into the following components: ‘space’, managing the virtual memory space and mapping objects, ‘thread_mgr’, which manages all threads owned by the monitor (typically, all threads started by compartments), ‘compartments’, which manages compartment state, and ‘dynlink’, which contains the dynamic linker state. The unmapper allows for background unmapping and cleanup of objects and handles. There are also two hangle managers, for the monitor to hand out handles to libraries and compartments to callers.

Fields
locks: LockCollection<(&'static RwLock<ThreadMgr>, &'static RwLock<CompartmentMgr>, &'static RwLock<&'static mut Context>, &'static RwLock<HandleMgr<LibraryHandle>>, &'static RwLock<HandleMgr<CompartmentHandle>>)>unmapper: OnceLock<Unmapper>space: &'static Mutex<Space>

Management of address space.

thread_mgr: &'static RwLock<ThreadMgr>

Management of all threads.

comp_mgr: &'static RwLock<CompartmentMgr>

Management of compartments.

dynlink: &'static RwLock<&'static mut Context>

Dynamic linker state.

library_handles: &'static RwLock<HandleMgr<LibraryHandle>>

Open handles to libraries.

_compartment_handles: &'static RwLock<HandleMgr<CompartmentHandle>>

Open handles to compartments.

Implementations

impl Monitor

pub(crate) fn start_compartment( &self, instance: ObjID, args: &[&CStr], env: &[&CStr], mondebug: bool, suspend_on_start: bool, ) -> Result<(), TwzError>

impl Monitor

pub fn get_compartment_info( &self, instance: ObjID, thread: ObjID, desc: Option<Descriptor>, ) -> Result<CompartmentInfo, TwzError>

Get CompartmentInfo for this caller. Note that this will write to the compartment-thread’s simple buffer.

pub fn get_compartment_gate_address( &self, instance: ObjID, thread: ObjID, desc: Option<Descriptor>, name_len: usize, ) -> Result<usize, TwzError>

Get CompartmentInfo for this caller. Note that this will write to the compartment-thread’s simple buffer.

pub fn get_compartment_handle( &self, caller: ObjID, compartment: ObjID, ) -> Result<Descriptor, TwzError>

Open a compartment handle for this caller compartment.

pub fn lookup_compartment( &self, instance: ObjID, thread: ObjID, name_len: usize, ) -> Result<Descriptor, TwzError>

Open a compartment handle for this caller compartment.

pub fn compartment_wait( &self, caller: ObjID, desc: Option<Descriptor>, flags: u64, ) -> u64

pub fn get_compartment_deps( &self, caller: ObjID, desc: Option<Descriptor>, dep_n: usize, ) -> Result<Descriptor, TwzError>

Open a handle to the n’th dependency compartment of a given compartment.

pub fn get_compartment_thread_info( &self, caller: ObjID, desc: Option<Descriptor>, t_n: usize, ) -> Result<ThreadInfo, TwzError>

Get the n’th thread’s info from a compartment.

pub fn load_compartment( &self, caller: ObjID, thread: ObjID, name_len: usize, args_len: usize, env_len: usize, new_comp_flags: NewCompartmentFlags, ) -> Result<Descriptor, TwzError>

Load a new compartment with a root library ID, and return a compartment handle.

pub fn drop_compartment_handle(&self, caller: ObjID, desc: Descriptor)

Drop a compartment handle.

pub fn update_compartment_flags( &self, instance: ObjID, f: impl FnOnce(u64) -> Option<u64>, ) -> bool

pub fn load_compartment_flags(&self, instance: ObjID) -> u64

pub fn wait_for_compartment_state_change(&self, instance: ObjID, state: u64)

impl Monitor

pub fn get_library_info( &self, instance: ObjID, thread: ObjID, desc: Descriptor, ) -> Result<LibraryInfo, TwzError>

Get LibraryInfo for a given library handle. Note that this will write to the compartment-thread’s simple buffer.

pub fn get_library_handle( &self, caller: ObjID, comp: Option<Descriptor>, num: usize, ) -> Result<Descriptor, TwzError>

Open a handle to the n’th library for a compartment.

pub fn load_library( &self, _caller: ObjID, _id: ObjID, _comp: Option<Descriptor>, ) -> Result<Descriptor, TwzError>

Load a library in the given compartment.

pub fn drop_library_handle(&self, caller: ObjID, desc: Descriptor)

Drop a library handle.

impl Monitor

pub fn stats(&self) -> MonitorStats

impl Monitor

pub fn start_background_threads(&self)

Start the background threads for the monitor instance. Must be done only once the monitor has been initialized.

pub fn new(init: InitDynlinkContext) -> Self

Build a new monitor state from the initial dynamic linker context.

pub fn start_thread( &self, main: Box<dyn FnOnce()>, ) -> Result<Arc<ManagedThreadInner>, TwzError>

Start a managed monitor thread.

pub fn spawn_compartment_thread( &self, instance: ObjID, args: ThreadSpawnArgs, stack_ptr: usize, thread_ptr: usize, ) -> Result<ObjID, TwzError>

Spawn a thread into a given compartment, using initial thread arguments.

pub fn get_comp_config( &self, sctx: ObjID, ) -> Result<*const SharedCompConfig, TwzError>

Get the compartment config for the given compartment.

pub fn map_object( &self, sctx: ObjID, info: MapInfo, ) -> Result<Arc<MapHandleInner>, TwzError>

Map an object into a given compartment.

pub fn map_pair( &self, sctx: ObjID, info: MapInfo, info2: MapInfo, ) -> Result<(Arc<MapHandleInner>, Arc<MapHandleInner>), TwzError>

Map a pair of objects into a given compartment.

pub fn unmap_object(&self, sctx: ObjID, info: MapInfo)

Unmap an object from a given compartmen.

pub fn get_thread_simple_buffer( &self, sctx: ObjID, thread: ObjID, ) -> Result<ObjID, TwzError>

Get the object ID for this compartment-thread’s simple buffer.

pub fn _write_thread_simple_buffer( &self, sctx: ObjID, thread: ObjID, bytes: &[u8], ) -> Result<usize, TwzError>

Write bytes to this per-compartment thread’s simple buffer.

pub fn read_thread_simple_buffer( &self, sctx: ObjID, thread: ObjID, len: usize, ) -> Result<Vec<u8>, TwzError>

Read bytes from this per-compartment thread’s simple buffer.

pub fn comp_name(&self, id: ObjID) -> Result<String, TwzError>

Read the name of a compartment.

pub fn upcall_handle( &self, frame: &mut UpcallFrame, info: &UpcallData, ) -> Result<Option<ResumeFlags>, TwzError>

pub fn compartment_ctrl( &self, info: &GateCallInfo, cmd: MonitorCompControlCmd, ) -> Option<i32>

Perform a compartment control action on the calling compartment.

pub fn set_nameroot( &self, _info: &GateCallInfo, root: ObjID, ) -> Result<(), TwzError>