monitor

Module mon

Re-exports

pub use space::early_object_map;

Modules

compartment 🔒

library

space 🔒

stat

thread 🔒

Structs

Monitor

A security monitor instance. All monitor logic is implemented as methods for this type. We split the state into the following components: ‘space’, managing the virtual memory space and mapping objects, ‘thread_mgr’, which manages all threads owned by the monitor (typically, all threads started by compartments), ‘compartments’, which manages compartment state, and ‘dynlink’, which contains the dynamic linker state. The unmapper allows for background unmapping and cleanup of objects and handles. There are also two hangle managers, for the monitor to hand out handles to libraries and compartments to callers.

Statics

MONITOR 🔒

Functions

get_monitor

Get the monitor instance. Panics if called before first call to set_monitor.

set_monitor

Set the monitor instance. Can only be called once. Must be called before any call to get_monitor.

Type Aliases

MonitorLocks 🔒